work on collision and player alignement

+++ srcs/parsing/map.c
@@ -14,8 +14,8 @@
 
 void	ft_not_square(t_map *map)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
     t_map *temp;
 
 	i = 0;
@@ -35,8 +35,8 @@ void	ft_not_square(t_map *map)
 
 int	ft_verify_wall(t_map *map)
 {
-    size_t i;
-	size_t 	j;
+    int i;
+	int 	j;
     t_map *temp;
 
 	i = 0;
